feat(app/worker-profile/jobs): separate common logic to hooks
