+++ packages/apps/human-app/frontend/src/i18n/en.json
@@ -187,6 +187,7 @@
       }
     },
     "jobs": {
+      "successFullyAssignedJob": "Successfully assigned a job!",
       "errorFetchingData": "There was an error while fetching data, please try again",
       "availableJobs": "Available Jobs",
       "myJobs": "My Jobs",
