+++ packages/apps/human-app/frontend/src/pages/worker/jobs/components/available-jobs/available-jobs-table-mobile.tsx
@@ -2,7 +2,6 @@
 import { Grid, List, Paper, Stack } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import type { Dispatch, SetStateAction } from 'react';
-import { useQueryClient } from '@tanstack/react-query';
 import { ProfileListItem } from '@/pages/operator/profile/profile-list-item';
 import { colorPalette } from '@/styles/color-palette';
 import { Button } from '@/components/ui/button';
@@ -12,12 +11,10 @@ import { useJobsFilterStore } from '@/hooks/use-jobs-filter-store';
 import { Loader } from '@/components/ui/loader';
 import { Alert } from '@/components/ui/alert';
 import { shortenEscrowAddress } from '@/shared/helpers/shorten-escrow-address';
-import type { AvailableJobsSuccessResponse } from '@/api/servieces/worker/available-jobs-data';
 import { getNetworkName } from '@/smart-contracts/get-network-name';
-import { useProtectedLayoutNotification } from '@/hooks/use-protected-layout-notifications';
-import { wait } from '@/shared/helpers/wait';
-import { defaultErrorMessage } from '@/shared/helpers/default-error-message';
 import { useAssignJobMutation } from '@/api/servieces/worker/assign-job';
+import { useJobsNotifications } from '@/hooks/use-jobs-notifications';
+import { useAvailableJobsTableState } from '@/hooks/use-available-jobs-table-state';
 
 interface AvailableJobsTableMobileProps {
   setIsMobileFilterDrawerOpen: Dispatch<SetStateAction<boolean>>;
@@ -26,43 +23,17 @@ interface AvailableJobsTableMobileProps {
 export function AvailableJobsTableMobile({
   setIsMobileFilterDrawerOpen,
 }: AvailableJobsTableMobileProps) {
-  const { filterParams } = useJobsFilterStore();
-  const { setTopNotification, closeNotification } =
-    useProtectedLayoutNotification();
-
-  const onJobAssignmentSuccess = async () => {
-    setTopNotification({
-      content: 'Successfully assigned a job!',
-      type: 'success',
-    });
-    await wait(5000);
-    closeNotification();
-  };
-
-  const onJobAssignmentError = async (error: unknown) => {
-    setTopNotification({
-      content: defaultErrorMessage(error),
-      type: 'warning',
-    });
-    await wait(5000);
-    closeNotification();
-  };
-
-  const queryClient = useQueryClient();
+  const { onJobAssignmentError, onJobAssignmentSuccess } =
+    useJobsNotifications();
 
   const { mutate: assignJobMutation } = useAssignJobMutation({
     onSuccess: onJobAssignmentSuccess,
     onError: onJobAssignmentError,
   });
 
-  const availableJobsTableState = queryClient.getQueryState([
-    'availableJobs',
-    filterParams,
-  ]);
-  const queryData = queryClient.getQueryData<AvailableJobsSuccessResponse>([
-    'availableJobs',
-    filterParams,
-  ]);
+  const { availableJobsTableState, availableJobsTableQueryData } =
+    useAvailableJobsTableState();
+
   const { t } = useTranslation();
 
   const { setSearchEscrowAddress } = useJobsFilterStore();
@@ -104,9 +75,8 @@ export function AvailableJobsTableMobile({
           </Stack>
         ) : null}
         {availableJobsTableState?.status === 'success' &&
-        !availableJobsTableState.error &&
-        queryData?.results
-          ? queryData.results.map((d) => (
+        !availableJobsTableState.error
+          ? availableJobsTableQueryData.map((d) => (
               <Paper
                 key={crypto.randomUUID()}
                 sx={{
